+++ config/bootstrap.php
@@ -14,6 +14,7 @@
 use Cake\Cache\Cache;
 use Cake\Core\Configure;
 use Cake\Core\Plugin;
+use Cake\Database\Type;
 use Cake\I18n\Date;
 use Cake\I18n\FrozenDate;
 use Cake\I18n\FrozenTime;
@@ -151,3 +152,5 @@ FrozenTime::setToStringFormat($format);
 Time::setToStringFormat($format);
 
 require_once __DIR__ . DS . 'i18n_constants.php';
+
+Type::map('jsonEntity', 'MeCms\Database\Type\JsonEntityType');